refactor(run): decompose runservice into functional modules
Extract the monolithic RunService (915 lines) into smaller, focused modules
using functional composition:

- run-service.ts: Now a thin wrapper (~258 lines)
- build-context.ts: Main context building logic (~211 lines)
- resolvers/: Separate resolution functions for checkpoint, session, conversation
- environment/: Environment variable expansion
- utils/: Shared utilities (extractWorkingDir, calculateSessionHistoryPath)

Key changes:
- Move ConversationResolution interface to resolvers/types.ts
- Extract expandEnvironmentFromCompose to environment/expand-environment.ts
- Extract resolveCheckpoint, resolveSession, resolveDirectConversation
- Create build-context.ts with buildExecutionContext function
- RunService.buildExecutionContext now delegates to functional implementation

Benefits:
- Each module is independently testable
- Clear separation of concerns
- No circular dependencies
- Public API unchanged (backward compatible)

Closes #1285
