+++ turbo/apps/web/src/lib/run/environment/index.ts
@@ -1,0 +1,2 @@
+export { expandEnvironmentFromCompose } from "./expand-environment";
+// ExpandedEnvironmentResult type is internal to this module
