+++ turbo/apps/web/src/lib/run/utils/index.ts
@@ -1,0 +1,2 @@
+export { extractWorkingDir } from "./extract-working-dir";
+// calculateSessionHistoryPath is exported from run-service.ts for backward compatibility
